Add Tds support for :inner_lateral and :left_lateral (#284)
This commit updates the TDS adapter to support the :inner_lateral
and :left_lateral joins, which map to OUTER APPLY and CROSS APPLY,
respectively.  For both of these joins, we don't need an ON clause.
Therefore, the join_on/4 function headers have been updated to omit
it.

Lastly, these require that the join contents be wrapped within
brackets ().  So, the join_expr function is added to wrap provided
join expression in brackets for these joins only.